[dev] updated api calls (wordcloud, nlp)
